+++ instrumentation/opentelemetry-instrumentation-requests/tests/test_requests_integration.py
@@ -22,7 +22,7 @@ import opentelemetry.instrumentation.requests
 from opentelemetry import context, propagators, trace
 from opentelemetry.instrumentation.requests import RequestsInstrumentor
 from opentelemetry.sdk import resources
-from opentelemetry.test.mock_httptextformat import MockHTTPTextFormat
+from opentelemetry.test.mock_textmap import MockTextMapPropagator
 from opentelemetry.test.test_base import TestBase
 from opentelemetry.trace.status import StatusCanonicalCode
 
@@ -147,29 +147,46 @@ class RequestsIntegrationTestBase(abc.ABC):
 
         self.assert_span(num_spans=0)
 
+    def test_not_recording(self):
+        with mock.patch("opentelemetry.trace.INVALID_SPAN") as mock_span:
+            RequestsInstrumentor().uninstrument()
+            # original_tracer_provider returns a default tracer provider, which
+            # in turn will return an INVALID_SPAN, which is always not recording
+            RequestsInstrumentor().instrument(
+                tracer_provider=self.original_tracer_provider
+            )
+            mock_span.is_recording.return_value = False
+            result = self.perform_request(self.URL)
+            self.assertEqual(result.text, "Hello!")
+            self.assert_span(None, 0)
+            self.assertFalse(mock_span.is_recording())
+            self.assertTrue(mock_span.is_recording.called)
+            self.assertFalse(mock_span.set_attribute.called)
+            self.assertFalse(mock_span.set_status.called)
+
     def test_distributed_context(self):
-        previous_propagator = propagators.get_global_httptextformat()
+        previous_propagator = propagators.get_global_textmap()
         try:
-            propagators.set_global_httptextformat(MockHTTPTextFormat())
+            propagators.set_global_textmap(MockTextMapPropagator())
             result = self.perform_request(self.URL)
             self.assertEqual(result.text, "Hello!")
 
             span = self.assert_span()
 
             headers = dict(httpretty.last_request().headers)
-            self.assertIn(MockHTTPTextFormat.TRACE_ID_KEY, headers)
+            self.assertIn(MockTextMapPropagator.TRACE_ID_KEY, headers)
             self.assertEqual(
                 str(span.get_context().trace_id),
-                headers[MockHTTPTextFormat.TRACE_ID_KEY],
+                headers[MockTextMapPropagator.TRACE_ID_KEY],
             )
-            self.assertIn(MockHTTPTextFormat.SPAN_ID_KEY, headers)
+            self.assertIn(MockTextMapPropagator.SPAN_ID_KEY, headers)
             self.assertEqual(
                 str(span.get_context().span_id),
-                headers[MockHTTPTextFormat.SPAN_ID_KEY],
+                headers[MockTextMapPropagator.SPAN_ID_KEY],
             )
 
         finally:
-            propagators.set_global_httptextformat(previous_propagator)
+            propagators.set_global_textmap(previous_propagator)
 
     def test_span_callback(self):
         RequestsInstrumentor().uninstrument()