+++ opentelemetry-api/src/opentelemetry/context/__init__.py
@@ -17,7 +17,6 @@ import threading
 import typing
 from functools import wraps
 from os import environ
-from sys import version_info
 
 from pkg_resources import iter_entry_points
 
@@ -48,11 +47,7 @@ def _load_runtime_context(func: _F) -> _F:
             if _RUNTIME_CONTEXT is None:
                 # FIXME use a better implementation of a configuration manager to avoid having
                 # to get configuration values straight from environment variables
-                if version_info < (3, 5):
-                    # contextvars are not supported in 3.4, use thread-local storage
-                    default_context = "threadlocal_context"
-                else:
-                    default_context = "contextvars_context"
+                default_context = "contextvars_context"
 
                 configured_context = environ.get(
                     "OTEL_CONTEXT", default_context